pyproject.toml
[project]
name = "nonebot-plugin-nsfw"
version = "0.1.0"
description = "群聊 NSFW 图片检测插件，带有撤回、禁言等功能。使用 Safety Checker / Inception V3"
authors = [{ name = "iyume", email = "iyumelive@gmail.com" }]
requires-python = ">=3.8"
readme = "README.md"
license = { text = "Apache-2.0" }
dependencies = [
    "nonebot2>=2.0",
    "nonebot-adapter-onebot>=2.0",
    "Pillow>=8.4.0",
    "httpx",
]

[project.optional-dependencies]
safety-checker = ["safety-checker[sdhook]>=0.3", "numpy"]
nsfw-model = ["tensorflow>=2.2.0", "tensorflow-hub>=0.12", "numpy"]


[tool.pyright]
typeCheckingMode = "basic"
reportUnknownParameterType = true # cannot check return...

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"
